+++ db/migrate/20200721055745_create_participation_events.rb
@@ -1,8 +1,8 @@
 class CreateParticipationEvents < ActiveRecord::Migration[6.0]
   def change
     create_table :participation_events do |t|
-      t.integer :event_id
-      t.integer :user_id
+      t.bigint :event_id
+      t.bigint :user_id
       t.text :purpose
       t.text :achivement
