+++ db/migrate/20200721060319_create_tag_events.rb
@@ -1,8 +1,8 @@
 class CreateTagEvents < ActiveRecord::Migration[6.0]
   def change
     create_table :tag_events do |t|
-      t.integer :tag_id
-      t.integer :event_id
+      t.bigint :tag_id
+      t.bigint :event_id
 
       t.timestamps
     end
