+++ db/migrate/20200721061247_create_communities.rb
@@ -1,7 +1,7 @@
 class CreateCommunities < ActiveRecord::Migration[6.0]
   def change
     create_table :communities do |t|
-      t.integer :user_id
+      t.bigint :user_id
       t.string :name
       t.text :introduction
       t.boolean :is_private
